move api calls and data out
